enh: Rename HomeComponents WIP
