+++ ajcf-new-site/components/Home/HomeComponents/HomeCarouselComponents/HomeCarouselContainer.tsx
@@ -2,10 +2,10 @@ import { LeftOutlined, RightOutlined } from "@ant-design/icons";
 import { Carousel } from "antd";
 import React, { useRef } from "react";
 import { ArticlePreviewFragment } from "../../../../types/types";
-import styles from "./Carousel.module.scss";
-import { CarouselItem } from "./CarouselItem";
+import styles from "./HomeCarousel.module.scss";
+import { HomeCarouselItem } from "./HomeCarouselItem";
 
-export const CarouselContainer = ({ articles }: { articles: ArticlePreviewFragment[] }) => {
+export const HomeCarouselContainer = ({ articles }: { articles: ArticlePreviewFragment[] }) => {
   const slider = useRef() as any;
   return (
     <section className={styles.carouselSection}>
@@ -17,7 +17,7 @@ export const CarouselContainer = ({ articles }: { articles: ArticlePreviewFragme
         }}
       >
         {articles.map((article) => (
-          <CarouselItem article={article} key={article.id} />
+          <HomeCarouselItem article={article} key={article.id} />
         ))}
       </Carousel>
       <RightOutlined className={styles.rightArrow} onClick={() => slider.current.next()} />
