+++ ajcf-new-site/components/Home/HomeComponents/HomeEspacesComponents/HomeEspacesContainer.tsx
@@ -1,16 +1,16 @@
 import React from "react";
-import styles from "./Espaces.module.scss";
-import { EspacesDescriptions } from "./EspacesDescriptions";
+import styles from "./HomeEspaces.module.scss";
+import { HomeEspacesDescriptions } from "./HomeEspacesDescriptions";
 
 const titleContent = "S'exprimer à travers nos 4 espaces";
 const textContent =
   "Et has minim elitr intellegat. Mea aeterno eleifend antiopam ad, nam no suscipit quaerendum. At nam minimum ponderum. Est audiam animal molestiae te.";
-export const Espaces = () => {
+export const HomeEspacesContainer = () => {
   return (
     <section className={styles.espaces}>
       <h2>{titleContent}</h2>
       <p>{textContent}</p>
-      <EspacesDescriptions />
+      <HomeEspacesDescriptions />
     </section>
   );
 };
