+++ ajcf-new-site/components/Home/HomeContainer.tsx
@@ -2,18 +2,18 @@ import React from "react";
 import { ArticlePreviewFragment } from "../../types/types";
 import { Layout } from "../Layout/Layout";
 import { Antennes } from "./HomeComponents/AntennesComponents/Antennes";
-import { CarouselContainer } from "./HomeComponents/CarouselComponents/CarouselContainer";
+import { HomeCarouselContainer } from "./HomeComponents/HomeCarouselComponents/HomeCarouselContainer";
 import { Education } from "./HomeComponents/EducationComponents/Education";
-import { Espaces } from "./HomeComponents/EspacesComponents/Espaces";
+import { HomeEspacesContainer } from "./HomeComponents/HomeEspacesComponents/HomeEspacesContainer";
 import { Partners } from "./HomeComponents/PartnersComponents/Partners";
-import { Stats } from "./HomeComponents/StatsComponents/Stats";
+import { HomeStatsContainer } from "./HomeComponents/HomeStatsComponents/HomeStatsContainer";
 
 export const HomeContainer = ({ articles }: { articles: ArticlePreviewFragment[] }) => {
   return (
     <Layout>
-      <Espaces />
-      <CarouselContainer articles={articles} />
-      <Stats />
+      <HomeEspacesContainer />
+      <HomeCarouselContainer articles={articles} />
+      <HomeStatsContainer />
       <Partners />
       <Education />
       <Antennes />
